fix postun script error
Former-commit-id: 9bef4cc1f8bfed6584abf6f605a3c6ca9364ffa8

+++ rpm/SPEC/smarter.spec
@@ -169,7 +169,10 @@ chkconfig --del celeryd-edextract
 
 
 %postun
-userdel -rf celery > /dev/null 2>&1
+id celery > /dev/null 2>&1
+if [ $? == 0 ] ; then
+    userdel -rf celery > /dev/null 2>&1
+fi
 
 
 %changelog
